Blood, sweat and tears...Trying to get the Log4j2 AWS Lambda config integrated in my setup that uses the Spring Boot Starter Parent...
Ok, perhaps just a bad idea, but hell why did it have to be so hard and utterly unhelpful regarding error messages.

As an aside added a profile to make it easy to exclude the slow npm/ng build, since I had to do a lot of clean/install cycles including my parent hierarchy.