Integrated the API for the UI into the Lambda.
The skill could not be registered as a servlet in the special lambda wrapper, so will have to investigate alternatives for that.

Not yet verified the UI against this version, but some postman smoke tests seemed ok and highlighted some issues in behavior. Not sure if this is an API GW thing or the serverless wrapper.

For now just copied the code around and not yet attempted to extract the commonalities with the plain spring boot app.